pyproject.toml
[project]
name = "gpt-assistant-command-line"
version = "0.1.0"
description = "Command-line interface for ChatGPT Assistants"
authors = [{name = "Lyon Lay", email = "lay.lyon@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
keywords = ["cli", "command-line", "assistant", "openai", "gpt-3", "gpt-4", "llm", "chatgpt", "gpt-cli", "gpt-client"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "attrs==23.1.0",
    "black==23.1.0",
    "openai==1.3.8",
    "prompt-toolkit==3.0.41",
    "pytest==7.3.1",
    "PyYAML==6.0",
    "rich==13.7.0",
    "tiktoken==0.5.2",
    "tokenizers==0.15.0",
    "typing_extensions==4.5.0",
]

[project.urls]
"Homepage" = "https://github.com/kharvd/gpt-cli"

[project.scripts]
gpt_assistant = "gptcli.gpt:main"

[build-system]
requires = ["pip>=23.0.0", "setuptools>=58.0.0", "wheel"]
build-backend = "setuptools.build_meta"

